requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
pydantic[email]
python-dotenv
bcrypt
python-jose[cryptography]
python-multipart

